Move mocks into test_utils.py
